energy/core: share CustomerAccountRFIDForm implementation
