+++ apps/energy/admin/forms.py
@@ -3,23 +3,7 @@ from __future__ import annotations
 from django import forms
 from django.utils.translation import gettext_lazy as _
 
-from ..models import CustomerAccount
-
-
-class CustomerAccountRFIDForm(forms.ModelForm):
-    """Form for assigning existing RFIDs to a customer account."""
-
-    class Meta:
-        model = CustomerAccount.rfids.through
-        fields = ["rfid"]
-
-    def clean_rfid(self):
-        rfid = self.cleaned_data["rfid"]
-        if rfid.energy_accounts.exclude(pk=self.instance.customeraccount_id).exists():
-            raise forms.ValidationError(
-                "RFID is already assigned to another customer account"
-            )
-        return rfid
+from .forms_shared import CustomerAccountRFIDForm
 
 
 class OdooCustomerSearchForm(forms.Form):
