+++ apps/energy/admin/forms_shared.py
@@ -1,0 +1,19 @@
+from django import forms
+
+from ..models import CustomerAccount
+
+
+class CustomerAccountRFIDForm(forms.ModelForm):
+    """Form for assigning existing RFIDs to a customer account."""
+
+    class Meta:
+        model = CustomerAccount.rfids.through
+        fields = ["rfid"]
+
+    def clean_rfid(self):
+        rfid = self.cleaned_data["rfid"]
+        if rfid.energy_accounts.exclude(pk=self.instance.customeraccount_id).exists():
+            raise forms.ValidationError(
+                "RFID is already assigned to another customer account"
+            )
+        return rfid
